Moving Callbacks.hpp file
DDE-8
